Replacing Xbox Controller Support with Use Input Manager Axes
