--- Assets/HoloToolkit/Utilities/Scripts/Editor/InputManagerAxesWindow.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 6718d53d90f9e474997d202900c98e60
-timeCreated: 1506970860
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
